refactor: extract channels logic in webhook handler into "processInboundChannelMessage" function
